Modul 5 - Login system

+++ application/views/post/index.php
@@ -9,7 +9,9 @@
     <div class="row">
         <div class="col-md-4 d-flex justify-content-between">
             <h1>Artikel</h1>
-            <a href="<?= base_url() ?>post/tambah" class="btn btn-primary align-self-center">Tambah Post</a>
+            <?php if (logged_in()) : ?>
+                <a href="<?= base_url() ?>post/tambah" class="btn btn-primary align-self-center">Tambah Post</a>
+            <?php endif; ?>
         </div>
     </div>
     <!-- <hr> -->
@@ -26,10 +28,11 @@
                     <h3 class="text-truncate"><?= $post['judul']; ?></h3>
                     <p class="" style="-webkit-line-clamp:3; overflow:hidden; text-overflow:ellipsis; display: -webkit-box; -webkit-box-orient:vertical;"><?= $post['isi']; ?>
                     </p>
-                    <a role="button" href="" class="btn btn-primary">Lihat &raquo;</a>
-                    <a role="button" href="<?= base_url() ?>post/update/<?= $post['id_post'] ?>" class="btn btn-success">Update</a>
-                    <a href="<?= base_url() ?>post/hapus/<?= $post['id_post'] ?>" class="btn btn-danger" onclick="return confirm('Yakin ingin menghapus Post?')">Hapus</a>
-
+                    <?php if (logged_in()) : ?>
+                        <a role="button" href="" class="btn btn-primary">Lihat &raquo;</a>
+                        <a role="button" href="<?= base_url() ?>post/update/<?= $post['id_post'] ?>" class="btn btn-success">Update</a>
+                        <a href="<?= base_url() ?>post/hapus/<?= $post['id_post'] ?>" class="btn btn-danger" onclick="return confirm('Yakin ingin menghapus Post?')">Hapus</a>
+                    <?php endif; ?>
                     <hr>
                 </div>
             <?php endforeach; ?>
